Lookup Files Return Value
This needs to be cross-checked. I don't know whether this is expected to be a string at runtime. The JSDoc is marked as an array of strings, so I would assume that this case was meant to be an array. Or, it could be the other way around, and this would be incorrect to make this change. I simply am making this change to reflect the JSDoc annotation.