Suite ES6 Class (Initial Migration)
I haven't gone through to validate the usage of fields yet. I will do that right after this commit.

I tend to not format my code when doing refactors like these, right out of the gate, because it allows the diffs to be neater and clearer as to what I specifically made edits to. I am happy to hear if this is correct or incorrect to handle things. For me usually I tend to get confused when it's a wall of text just for a simple change. It's nice to see formatting spread out into a separate commit, in my opinion thus far.