Test ES6 Class (Initial Migration)
I think there are just a few fields to declare here unique to this class. Will follow that up with the next commit.

Note that I am moving the `module.exports = Test` (I am doing this for all ES6 class conversions) because `class` is not hoisted like `function` is, so you have to either export it on the same line, or export it after having declared it. I would leave the export in the same place if I were able to do so.

I also tend not to export right as it's declared (in a CJS context) because it removes it from the scope of the file, which I don't want. Sometimes adjacent references use it as it's name itself, and CJS export assignments don't allow for this, unlike `export class Test` would in ESM. If this were ESM now, I could do that, and it would still be accessible in the module scope, but ESM will be a separate migration. I've found that doing ESM separate from type checking seems to help a lot.