Stats Collector Optional Fields
I don't think these are necessarily intentional by design, at least that I'm aware? The only thing I can think of is that they are not present at all times simply because of when they are appended to the object. Will have to discover this later with time, and more type inference later down the road.

I am marking currently "complete" files with `// @ts-nocheck` so I can find the files I need to adjust next. Then I will do another pass to find the other errors which need addressing afterwards.

This first pass of typings is simply to fix and cover what we already have. Some inferences are fixed simply by when other files are addressed, so when something appears to be incorrect coming from a different module, I don't go try to fix it right out of the gate, I simply mark that current file as "done" with a nocheck flag, then continue fixing things I can currently fix with the current state of the codebase, going up the module graph, rather than getting sidetracked by issues coming from features present in other modules.

Say like `test.cases` is not present. Hypothetically I am in `suite.js`. `Test` is defined in `test.js`, so I don't need to fix that error inside of `suite.js` yet, because it revolves on the fix of an issue over in `test.js`.