Context ES6 Class
Not many unchecked errors at this point! May start moving over to ESM after this.

There are still plenty of errors, but ones out of scope of the first wave of fixes, to clarify. It's a lot of scoping/usage issues, like if a class's method or field is marked as private, that kind of thing.