Mocha (root module) ESM
Note that this one is a bit more complex than the others, so there are still a few cases where `require()` calls haven't been ported yet. I'm not sure how those are intended to be solved, whether that's with the require builder feature, or dynamic imports.

The `module.paths` feature hasn't been accounted for either, I am not totally familiar with that enough to know how to fix it yet. So I removed the `module.` prefix so it can temporarily show as a type error to be fixed.

One of the big things with porting code when using type checking, is that `require()` and `module.exports`, `module`, and `exports` don't get flagged as a compiler type error when accessed in ESM. They absolutely will be a runtime error though, so that's not good. I really worked to make sure that no dangling uses of `require()` or the `module.exports` variations were left behind anywhere. This seems to be the module with the most exceptions to that. This one will need a good deal more work.