Type Checking Changes Revert
This is because I seemed to have started to add type checking, fixed the errors, then migrated the ES5 constructor functions over to ES6 classes. But these ideally should be separate commits/concerns, so this branch is intended to be solely just the class conversion code changes, not type checking and or ESM related changes.