Additional Checking Revert
Streamlining some of the parts of the code so the diffs are clearer (that docs didn't change, and they were simply moved to other parts of the file), and removing additional type annotations which aren't necessary at this stage of the migration.

I also removed references to `util.inherits()` as that is the goal with this migration too. Some modules still use a namespace for the import though, so for those I didn't change anything.

This and the previous commit are meant to be additive changes to make the diff from my changes the last few days into something more streamlined and intentional. They make more sense if you were to diff everything together, which is what I did to make these changes themselves.