Static Fields Revert
This seemingly is a 2022 feature, which forgot was relatively new. Static methods have been around since ES2015, while static fields are in ES2022. We seem to not be using a parser which will allow this syntax, as it gets caught up on the `=` usage, when running `npm run test`, and `npm run build`.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Classes/static