Base Reporter ES6 Class Revert
Because the module defines more than one field on the main class which is exported from the file, this broke with my changes of when I assigned `exports` and `module.exports` at the end of the file. For making the PR simpler and streamlining which features do and don't work, I am going to refactor this to work so there are no changes to the test results.

On the main branch (8649f39), this is the results of the test suite:

  350 passing (2m)
  1 pending
  2 failing

With this commit, the tests still aren't symmetrical yet. There is one other file, which I will revert separate to this one, for clarity as well.

Again, this brings this file back to the state it was in on that main branch commit mentioned above. I will rework this file so that my changes will not error like this.