Mocha ES6 Class Revert
Further info related to this in the previous commit.

This revert brings this file back to the state of the main branch before adding any of my own changes to the project. With this commit, now the test suite is symmetrical to how it was before I made my changes to the project.