XUnit Call Syntax Fallback
This is a fallback to allow for calling an ES6 class with `.call`, and not `new`. This is simply to fix issues with certain tests not passing as expected. Ideally the code for those will migrate to using `new`, but I'm not sure whether changing the tests are an option.