Runnable globalArgs Secondary Fix
This required a full rename at the editor level, I had accepted this rename over on GitHub, but it only changed the parameter name, not the rest of the function body/type definitions. I think my error was that I took the issue as the full fix, rather than a suggestion of things to edit fully on my own, in the editor (rename symbol, that kind of thing).